Split base/identifier definitions from parsing.
